core/gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
redis==5.0.7
python-dotenv==1.0.1
sentence-transformers==3.0.1
psutil==5.9.8
docker==7.0.0
websockets==12.0
pynvml==11.5.0
GPUtil==1.4.0
psycopg2-binary==2.9.9
pyyaml==6.0.1